Reflow static content on terminal size change
This change basically reimplements a big chunk of the Ink renderer and
output logic.  It gets the Ink core out of the business of caring
about screen refreshes and terminal size.  It also gets the index
count out of the Static component.  Instead, we make rendering
idempotent and move the logic that makes output characters dependent
on terminal state into the renderer.  Instead of deleting things
incrementally from the static component, we render every static
component and rely on logic in the bowels of the renderer to avoid
doing full screen refreshes when possible.

The resulting model is simpler, more React-y, and adjusts better to
terminal size.

We also by side effect address an unbounded memory use bug in the old
renderer: because the old renderer accumulated static content in
ink.tsx directly and output it outside React's normal refresh flow,
our static content history grew forever.  In long sessions, with lots
of /refresh commands, it could reach tens of thousands of lines.

Now, our entire history is encapsulated in the static component, and
when we clear or remount that, we forget obsolete data and keep our
memory use (and terminal spam) bounded as long as the size of the
static is bounded.